requirements.txt
# JUMP0X1 Dependencies
# Install with: pip install -r requirements.txt

# HTTP/WebSocket
requests>=2.28.0
aiohttp>=3.8.0

# Polymarket CLOB Client
py-clob-client>=0.14.0

# Environment variables
python-dotenv>=1.0.0
